Updated Front end feature

+++ app/Http/Controllers/Admin/ProductController.php
@@ -27,7 +27,8 @@ class ProductController extends Controller
         $this->middleware('permission:view-product')->only('index');
     }
 
-    public function index(ProductsDataTable $dataTable) {
+    public function index(ProductsDataTable $dataTable)
+    {
 
         $categories = Category::with('children')->whereNull('parent_id')->get();
         return $dataTable->render('admin.products.index', [
@@ -35,9 +36,9 @@ class ProductController extends Controller
         ]);
     }
 
-    public function show($id, ProductSkusDataTable $dataTable)
+    public function show($slug, ProductSkusDataTable $dataTable)
     {
-        $product = Product::find($id);
+        $product = Product::where('slug', $slug)->firstOrFail();
         $product_attributes = $product->product_attribute_set->attributes;
         $category = $product->categories()->where('id', $product->category_id)->first();
         return $dataTable->with(['product' => $product])->render('admin.products.show', [
@@ -59,43 +60,45 @@ class ProductController extends Controller
     public function store(Request $request)
     {
         $description = $request->input('description');
-        $description = substr($description, 3, strlen($description) - 7);
         $request->validate([
 //            'product_image' => 'required|image|mimes:jpeg,png,jpg|max:5048',
             'title' => 'required|unique:products',
         ]);
-        if($request->filled('filepath')) {
+        if ($request->filled('filepath')) {
             $image_path = $request->input('filepath');
             $image_path = explode('http://localhost:8000', $image_path)[1];
-        }else {
+        } else {
             $image_path = '/storage/photos/products/empty-photo.jpg';
         }
 
 
-        if($request->filled('filepath_1')) {
+        if ($request->filled('filepath_1')) {
             $image_path_1 = $request->input('filepath_1');
             $image_path_1 = explode('http://localhost:8000', $image_path_1)[1];
-        }else {
+        } else {
             $image_path_1 = '/storage/photos/products/empty-photo.jpg';
         }
-        if($request->filled('filepath_2')) {
+        if ($request->filled('filepath_2')) {
             $image_path_2 = $request->input('filepath_2');
             $image_path_2 = explode('http://localhost:8000', $image_path_2)[1];
-        }else {
+        } else {
             $image_path_2 = '/storage/photos/products/empty-photo.jpg';
         }
-        if($request->filled('filepath_3')) {
+        if ($request->filled('filepath_3')) {
             $image_path_3 = $request->input('filepath_3');
             $image_path_3 = explode('http://localhost:8000', $image_path_3)[1];
-        }else {
+        } else {
             $image_path_3 = '/storage/photos/products/empty-photo.jpg';
         }
 
         $product = Product::create([
             'title' => $request->input('title'),
             'price' => $request->input('price'),
+            'price_old' => $request->input('price_old'),
+            'total_item' => $request->input('total_item'),
             'percent_sale' => $request->percent_sale,
             'description' => $description,
+            'content' => $request->input('content'),
             'image' => $image_path,
             'category_id' => $request->category_id,
             'status' => 1,
@@ -115,11 +118,10 @@ class ProductController extends Controller
         return redirect(ProductController::PRODUCTS_PATH);
     }
 
-    public function edit($id)
+    public function edit($slug)
     {
         $product_attribute_sets = Product_Attribute_Set::select('id', 'name')->get();
-        $product = Product::find($id);
-
+        $product = Product::where('slug', $slug)->first();
         $product_attribute_set = $product->product_attribute_set;
 
         $category = Category::find($product->category_id);
@@ -145,48 +147,53 @@ class ProductController extends Controller
             ]);
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, $slug)
     {
         $description = $request->input('description');
-        $description = substr($description, 3, strlen($description) - 7);
 
-        $product = Product::find($id);
+        $product = Product::where('slug', $slug)->first();
+        $array_product_images = $product->product_images->pluck('image_url');
 
-        if($request->filled('filepath')) {
-            $image_path = $request->input('filepath');
-            $image_path = explode('http://localhost:8000', $image_path)[1];
-        }else {
-            $image_path = $product->image;
-        }
+        if ($product) {
+            if ($request->filled('filepath')) {
+                $image_path = $request->input('filepath');
+                $image_path = explode('http://localhost:8000', $image_path)[1];
+            } else {
+                $image_path = $product->image;
+            }
 
-        // sub image
-        if($request->filled('filepath_1')) {
-            $image_path_1 = $request->input('filepath_1');
-            $image_path_1 = explode('http://localhost:8000', $image_path_1)[1];
-        }else {
-            $image_path_1 = $product->image;
-        }
-        if($request->filled('filepath_2')) {
-            $image_path_2 = $request->input('filepath_2');
-            $image_path_2 = explode('http://localhost:8000', $image_path_2)[1];
-        }else {
-            $image_path_2 = $product->image;
-        }
-        if($request->filled('filepath_3')) {
-            $image_path_3 = $request->input('filepath_3');
-            $image_path_3 = explode('http://localhost:8000', $image_path_3)[1];
-        }else {
-            $image_path_3 = $product->image;
+            // sub image
+            if ($request->filled('filepath_1')) {
+                $image_path_1 = $request->input('filepath_1');
+                $image_path_1 = explode('http://localhost:8000', $image_path_1)[1];
+            } else {
+                $image_path_1 = $array_product_images[0];
+            }
+            if ($request->filled('filepath_2')) {
+                $image_path_2 = $request->input('filepath_2');
+                $image_path_2 = explode('http://localhost:8000', $image_path_2)[1];
+            } else {
+                $image_path_2 = $array_product_images[1];
+            }
+            if ($request->filled('filepath_3')) {
+                $image_path_3 = $request->input('filepath_3');
+                $image_path_3 = explode('http://localhost:8000', $image_path_3)[1];
+            } else {
+                $image_path_3 = $array_product_images[2];
+            }
         }
 
         $product->update([
             'title' => $request->input('title'),
             'price' => $request->input('price'),
+            'price_old' => $request->input('price_old'),
             'percent_sale' => $request->percent_sale,
             'description' => $description,
+            'content' => $request->input('content'),
             'image' => $image_path,
             'category_id' => $request->category_id,
             'status' => $request->input('status'),
+            'total_item' => $request->input('total_item'),
             'deleted_at' => null,
             'product_attribute_set_id' => $request->input('product_attribute_set_id'),
         ]);
@@ -211,26 +218,29 @@ class ProductController extends Controller
                 array_shift($image_paths);
             }
         }
-
-
         return redirect('/admin/products');
     }
 
-    public function destroy($id, Request $request)
+    public function destroy($slug, Request $request)
     {
-        if($request->filled('id')) {
-            Product::find($id)->delete();
+        if ($request->filled('slug')) {
+            $product = Product::where('slug', $slug)->first();
+            if ($product) {
+                $product->delete();
+                $product_images = Product_Image::where('product_id', $product->id)->get();
+
+                foreach ($product_images as $product_image) {
+                    $product_image->delete();
+                }
+            } else {
+                echo "Product not found.";
+            }
         }
 
-        $product_images = Product_Image::where('product_id', $id)->get();
-
-        foreach ($product_images as $product_image) {
-            $product_image->delete();
-        }
 
         $model = Product::query()
-            ->select(['id', 'image', 'title', 'description', 'price', 'status', 'created_at', 'updated_at'])
-            ->where('deleted_at','<>', 'null')
+            ->select(['id', 'image', 'title', 'description', 'price', 'status', 'created_at', 'updated_at', 'price_old', 'percent_sale'])
+            ->where('deleted_at', '<>', 'null')
             ->where('status', 4);
 
         return DataTables::of($model)
@@ -262,11 +272,11 @@ class ProductController extends Controller
     public function softDelete(Request $request)
     {
         $model = Product::query()
-            ->select(['id', 'image', 'title', 'description', 'price', 'status', 'created_at', 'updated_at'])
+            ->select(['id', 'image', 'title', 'description', 'price', 'status', 'created_at', 'updated_at', 'slug', 'price_old', 'percent_sale'])
             ->whereNull('deleted_at')
             ->where('status', '<>', 4);
-        if ($request->filled('product_id')) {
-            Product::find($request->product_id)->update([
+        if ($request->filled('slug')) {
+            Product::where('slug', $request->slug)->first()->update([
                 'deleted_at' => now(),
                 'status' => 4
             ]);
@@ -295,39 +305,6 @@ class ProductController extends Controller
             ->rawColumns(['image'])
             ->make();
     }
-
-    public function trashProduct()
-    {
-        $model = Product::query()
-            ->select(['id', 'image', 'title', 'description', 'price', 'status', 'created_at', 'updated_at'])
-            ->where('deleted_at','<>', 'null')
-            ->where('status', 4);
-
-        return DataTables::of($model)
-            ->editColumn('status', function ($product) {
-                if ($product->status == 1) {
-                    return 'Hoạt động';
-                } elseif ($product->status == 2) {
-                    return 'Không hoạt động';
-                } elseif ($product->status == 3) {
-                    return 'Đợi';
-                } else {
-                    return 'Xóa mềm';
-                }
-            })
-            ->editColumn('price', function ($product) {
-                return number_format($product->price * 1000, 0, ',', ',');
-            })
-            ->addColumn('action', function ($product) {
-                return view('admin.products.action_delete', ['product' => $product]);
-            })
-            ->editColumn('image', function ($row) {
-                return '<img class="img-thumbnail user-image-45" src="' . $row->image . '" alt="' . $row->title . '">';
-            })
-            ->rawColumns(['image'])
-            ->make();
-    }
-
 
     public function import()
     {
